UPDATE : 재사용 가능한 테스트 유틸리티 추가, gradle 스크립트 수정.
issue : https://github.com/JustBurrow/kobalttown/issues/9
time : 24h 45m

테스트 소스로 의존성 설정하기엔 문제가 많은 듯.
