+++ account-dao/build.gradle
@@ -2,7 +2,7 @@ dependencies {
   api project(':account-domain')
   implementation project(':account-jpa')
 
-  testCompile project(path: ':account-jpa', configuration: 'testArchives')
-  testCompile project(':configuration-jpa')
-  testCompile 'mysql:mysql-connector-java'
+  testImplementation project(':test-util')
+  testImplementation project(':configuration-jpa')
+  testImplementation 'mysql:mysql-connector-java'
 }
