--- account-dao/src/test/java/test/kr/lul/kobalttown/account/dao/AccountDaoTestUtil.java
@@ -1,38 +1,0 @@
-package test.kr.lul.kobalttown.account.dao;
-
-import kr.lul.kobalttown.account.dao.AccountDao;
-import kr.lul.kobalttown.account.domain.Account;
-import kr.lul.kobalttown.account.domain.Credential;
-import kr.lul.kobalttown.account.jpa.entity.AccountEntity;
-import org.slf4j.Logger;
-import org.springframework.beans.factory.annotation.Autowired;
-import test.kr.lul.kobalttown.account.jpa.AccountJpaTestUtil;
-
-import javax.annotation.PostConstruct;
-
-import static org.assertj.core.api.Assertions.assertThat;
-import static org.slf4j.LoggerFactory.getLogger;
-
-/**
- * @author justburrow
- * @since 2019-04-10
- */
-public class AccountDaoTestUtil extends AccountJpaTestUtil {
-  private static final Logger log = getLogger(AccountDaoTestUtil.class);
-
-  @Autowired
-  private AccountDao accountDao;
-
-  @PostConstruct
-  private void postConstruct() {
-    assertThat(this.accountDao).isNotNull();
-  }
-
-  public Account createdAccount() {
-    return this.accountDao.create(prePersistAccount());
-  }
-
-  public Credential createdCredential() {
-    return this.accountDao.create(prePersistCredential((AccountEntity) createdAccount()));
-  }
-}
