+++ account-jpa/src/test/java/kr/lul/kobalttown/account/jpa/repository/AccountRepositoryTest.java
@@ -2,6 +2,7 @@ package kr.lul.kobalttown.account.jpa.repository;
 
 import kr.lul.kobalttown.account.domain.Account;
 import kr.lul.kobalttown.account.jpa.entity.AccountEntity;
+import kr.lul.kobalttown.test.account.AccountJpaTestUtil;
 import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
@@ -10,8 +11,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
 import org.springframework.test.context.ContextConfiguration;
 import org.springframework.test.context.junit4.SpringRunner;
-import test.kr.lul.kobalttown.account.jpa.AccountJpaTestConfiguration;
-import test.kr.lul.kobalttown.account.jpa.AccountJpaTestUtil;
+import test.configuration.AccountJpaTestConfiguration;
 
 import java.time.Instant;
 import java.util.List;
