+++ article-dao/build.gradle
@@ -3,7 +3,7 @@ dependencies {
   implementation project(':article-jpa')
   api project(':support-spring-boot')
 
-  testImplementation project(path: ':article-jpa', configuration: 'testArchives')
+  testImplementation project(':test-util')
   testImplementation project(':configuration-jpa')
   testImplementation 'mysql:mysql-connector-java'
 }
