+++ article-dao/src/test/java/kr/lul/kobalttown/article/dao/ArticleDaoImplTest.java
@@ -9,7 +9,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.test.context.junit4.SpringRunner;
 import org.springframework.transaction.annotation.Transactional;
-import test.kr.lul.kobalttown.article.dao.ArticleDaoTestConfiguration;
+import test.configuration.ArticleDaoTestConfiguration;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
