+++ article-jpa/build.gradle
@@ -3,8 +3,7 @@ dependencies {
   implementation project(':account-jpa')
   api project(':support-spring-jpa')
 
-  testCompile project(path: ':article-domain', configuration: 'testArchives')
-  testCompile project(path: ':account-jpa', configuration: 'testArchives')
-  testCompile project(':configuration-jpa')
+  testImplementation project(':test-util')
+  testImplementation project(':configuration-jpa')
   testImplementation 'com.h2database:h2'
 }
