--- article-jpa/src/test/java/test/kr/lul/kobalttown/article/jpa/ArticleJpaTestConfiguration.java
@@ -1,42 +1,0 @@
-package test.kr.lul.kobalttown.article.jpa;
-
-import kr.lul.kobalttown.article.jpa.ArticleJpaConfiguration;
-import kr.lul.kobalttown.common.util.SystemTimeProvider;
-import kr.lul.kobalttown.common.util.TimeProvider;
-import kr.lul.kobalttown.configuration.jpa.ConfigurationJpaConfiguration;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.boot.autoconfigure.domain.EntityScan;
-import org.springframework.context.annotation.Bean;
-import org.springframework.data.jpa.repository.config.EnableJpaRepositories;
-import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
-import org.springframework.security.crypto.password.PasswordEncoder;
-import test.kr.lul.kobalttown.account.jpa.AccountJpaTestUtil;
-
-/**
- * @author justburrow
- * @since 2019-04-01
- */
-@SpringBootApplication(scanBasePackageClasses = {ConfigurationJpaConfiguration.class, ArticleJpaConfiguration.class})
-@EnableJpaRepositories(basePackages = ConfigurationJpaConfiguration.JPA_BASE_SCAN_PACKAGE)
-@EntityScan(basePackages = ConfigurationJpaConfiguration.JPA_BASE_SCAN_PACKAGE)
-public class ArticleJpaTestConfiguration {
-  @Bean
-  public TimeProvider timeProvider() {
-    return new SystemTimeProvider();
-  }
-
-  @Bean
-  public PasswordEncoder passwordEncoder() {
-    return new BCryptPasswordEncoder();
-  }
-
-  @Bean
-  public AccountJpaTestUtil accountJpaTestUtil() {
-    return new AccountJpaTestUtil();
-  }
-
-  @Bean
-  public ArticleJpaTestUtil articleJpaTestUtil() {
-    return new ArticleJpaTestUtil();
-  }
-}
