+++ runner-standalone/src/main/java/kr/lul/kobalttown/runner/standalone/StandaloneRunner.java
@@ -1,7 +1,6 @@
 package kr.lul.kobalttown.runner.standalone;
 
 import kr.lul.kobalttown.account.web.AccountWebConfiguration;
-import kr.lul.kobalttown.configuration.bean.ConfigurationBeanConfiguration;
 import kr.lul.kobalttown.configuration.jpa.ConfigurationJpaConfiguration;
 import kr.lul.kobalttown.configuration.web.ConfigurationWebConfiguration;
 import kr.lul.kobalttown.root.web.RootWebConfiguration;
@@ -15,8 +14,8 @@ import static org.slf4j.LoggerFactory.getLogger;
  * @author justburrow
  * @since 2019-02-27
  */
-@SpringBootApplication(scanBasePackageClasses = {
-    ConfigurationBeanConfiguration.class, ConfigurationWebConfiguration.class, ConfigurationJpaConfiguration.class,
+@SpringBootApplication(scanBasePackageClasses = {StandaloneRunnerConfiguration.class,
+    ConfigurationWebConfiguration.class, ConfigurationJpaConfiguration.class,
     AccountWebConfiguration.class, RootWebConfiguration.class})
 public class StandaloneRunner {
   private static final Logger log = getLogger(StandaloneRunner.class);
