+++ runner-standalone/src/main/java/kr/lul/kobalttown/runner/standalone/StandaloneRunnerConfiguration.java
@@ -1,0 +1,18 @@
+package kr.lul.kobalttown.runner.standalone;
+
+import kr.lul.kobalttown.common.util.SystemTimeProvider;
+import kr.lul.kobalttown.common.util.TimeProvider;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+
+/**
+ * @author justburrow
+ * @since 2019-04-19
+ */
+@Configuration
+public class StandaloneRunnerConfiguration {
+  @Bean
+  public TimeProvider timeProvider() {
+    return new SystemTimeProvider();
+  }
+}
